コード（関数）を見やすく10

+++ Engine/SEAHUNTER/PlayGame/Block.cpp
@@ -17,7 +17,7 @@ void Block::Initialize(int type, XMFLOAT3 pos)
 	{
 		rockBlock_ = Object3d::Create(ObjFactory::GetInstance()->GetModel("Rock"));
 		rockBlock_->SetPosition({ pos.x, 0.8f + pos.y, pos.z });
-		rockBlock_->SetRotation({ RandCalculate(0.0f,180.0f), 0,RandCalculate(0.0f,180.0f) });
+		rockBlock_->SetRotation({ RandCalculate(0.0f,180.0f), 0.0f, RandCalculate(0.0f,180.0f) });
 		float size = RandCalculate(1.0f, 3.0f);
 		rockBlock_->SetScale({ size, size, size });
 
@@ -25,18 +25,18 @@ void Block::Initialize(int type, XMFLOAT3 pos)
 	}
 	else if (type == (int)BlockType::Coral)
 	{
-		int numCount = RandCalculate(3.0f, 5.0f);
+		int numCount = (int)RandCalculate(3.0f, 5.0f);
 
 		for (int i = 0; i < numCount; i++)
 		{
 			CoralData tmp;
 			tmp.coralBlock = Object3d::Create(ObjFactory::GetInstance()->GetModel("coral"));
 
-			float size = RandCalculate(1.0f, 5.0f) / 10;
+			float size = RandCalculate(1.0f, 5.0f) / 10.0f;
 			tmp.coralBlock->SetScale({ size, size, size });
 
 			tmp.coralBlock->SetPosition({ pos.x + RandCalculate(-size,size), size + pos.y, pos.z + RandCalculate(-size,size) });
-			tmp.coralBlock->SetRotation({ 0, RandCalculate(0.0f,180.0f),0 });
+			tmp.coralBlock->SetRotation({ 0.0f, RandCalculate(0.0f,180.0f), 0.0f });
 
 			tmp.bubbleParticle = std::make_unique<ObjParticle>();
 			tmp.bubbleEmitter = std::make_unique<ParticleEmitter>(tmp.bubbleParticle.get());
@@ -44,8 +44,8 @@ void Block::Initialize(int type, XMFLOAT3 pos)
 			tmp.bubbleEmitter->SetCenter(size);
 			tmp.bubbleEmitter->SetObjStartScale({ scale, scale, scale });
 			tmp.bubbleEmitter->SetObjEndScale({ scale, scale, scale });
-			tmp.bubbleEmitter->SetStartColor({ 1,1,1,0.5f });
-			tmp.bubbleEmitter->SetEndColor({ 1,1,1,0.2f });
+			tmp.bubbleEmitter->SetStartColor({ 1.0f, 1.0f, 1.0f, 0.5f });
+			tmp.bubbleEmitter->SetEndColor({ 1.0f, 1.0f, 1.0f, 0.2f });
 
 			coralBlock_.push_back(std::move(tmp));
 		}
